Ändern des Datentyps von int zu float beim Plotter Datentyp für belegten Speicher und Veränderung jetzt float -> Mittelwertdatei speichert belegten Specher als float -> Umwandlung von float in int funktioniert nicht gut -> Keine Nachteile für Plotter, wenn Datentyp float ist

+++ Anwendungsfaelle/test_anhaengen/plotter_speicher_belegt.py
@@ -11,8 +11,8 @@ dateiinhalt: list = [zeile.split(",") for zeile in dateiinhalt if zeile]
 
 # Sortieren der benoetigten Daten in seperate Listen
 wiederholung: list = [int(zeile[0]) for zeile in dateiinhalt[1:]]
-speicher_liste_gesamt: list = [int(zeile[1]) for zeile in dateiinhalt[1:]]
-speicher_array_gesamt: list = [int(zeile[2]) for zeile in dateiinhalt[1:]]
+speicher_liste_gesamt: list = [float(zeile[1]) for zeile in dateiinhalt[1:]]
+speicher_array_gesamt: list = [float(zeile[2]) for zeile in dateiinhalt[1:]]
 
 # Grafische Darstellung der Listen
 liste, = plt.plot(wiederholung, speicher_liste_gesamt, label="Liste")
